Use floats for stations scores, change station totals

+++ db/migrate/20120510231910_create_people.rb
@@ -3,9 +3,9 @@ class CreatePeople < ActiveRecord::Migration
     create_table :people do |t|
       t.string :first_name, :null => false, :default => "John"
       t.string :last_name, :null => false, :default => "Doe"
-      t.integer :practical_score, :null => false, :default => 0
-      t.integer :written_score, :null => false, :default => 0
-      t.integer :communication_score, :null => false, :default => 0
+      t.float :practical_score, :null => false, :default => 0
+      t.float :written_score, :null => false, :default => 0
+      t.float :communication_score, :null => false, :default => 0
       t.integer :status, :null => false, :default => 0
     end
   end
